+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### unreleased
 
+* enhancements
+  * Add support for Rails 7.0 alpha 2.
+
 ### 4.8.0 - 2021-04-29
 
 * enhancements